Refactor the type erasure in IManager
Move ManagerModel to its own header file so it is not distributed with
IManager so that the partial interface is enforced by a file that the
plugin developer is unable to edit.

Refactor IManager and PluginManager to pass around ManagerConcept*
instead of letting IManager create a new ManagerModel since it no longer
has knowledge of that class.