Fix versioning to be checked by PluginManager
The versioning is no longer enforced by the plugin (since a plugin could
previously just return true and the PluginManager would trust it).

Plugins now contain a version number for the PluginManager they are
targeting -- must be provided by the application developer for the
plugin developers.

PluginManager now contains MIN_VERSION and MAX_VERSION which are the
versions of the PluginManager that it is compatible with, and are used
in the initialization of plugins to ensure a plugin is compatible before
loading it.