Remove #defines from plugin_manager.h
Remove the use of #defines in plugin_manager.h and prefer const float
values that are a property of the PluginManager class.

Add an undef directive in imanager.h to clean up the #define for VERSION
in case it should be defined elsewhere -- this is to avoid the use of a
global variable unnecessarily, and allow the plugin developer to easily
change what the target version for the plugin manager.